bugfix: Delete *output_brigade buckets after we've sent them to the FastCGI IPC socket.
If the request body is large enough, it is saved to a temporary file.
Looping through FILE buckets morphs them into numerous small HEAP buckets, and
the Apache child effectively allocates as much memory as the total size
of the *output_brigade data. This could be hundreds of MBytes, and
is the main reason why we saved the request body in a temp file, and not
directly in memory.

Deleting the buckets while we loop through *output_brigade frees
the memory used by the splitted and morphed HEAP buckets. This
method of iterating is recommended by the APR library docs.

The *output_brigade is later not used in other functions.
The current function proc_write_ipc() is its final destination.